Rename default_gas_costs to default_gas_costs_for_tests
