+++ fuel-tx/src/transaction/consensus_parameters/gas/default_gas_costs_for_tests.rs
